Basic Melee Script and Test in UItests 1
A basic melee script with a test run of it in the UItests 1. I added comments to the scripts descripting how each function works. This shouldn't have any problems. Made sure to make the changes of the player prefab in the scene and not to the actual prefab in the assets folder. Added a new layer as well for the Physics2d.OverlapCircleAll function.